--- src/components/molecules/card/CardFooter.tsx
@@ -1,10 +1,0 @@
-import { cn } from '@/shared/utils/cn'
-
-export function CardFooter({
-  className,
-  ...props
-}: React.ComponentProps<'div'>) {
-  return (
-    <div data-slot="card-footer" className={cn('flex', className)} {...props} />
-  )
-}
