--- src/components/molecules/card/CardHeader.tsx
@@ -1,10 +1,0 @@
-import { cn } from '@/shared/utils/cn'
-
-export function CardHeader({
-  className,
-  ...props
-}: React.ComponentProps<'div'>) {
-  return (
-    <div data-slot="card-header" className={cn('flex', className)} {...props} />
-  )
-}
